[TableGen] Yield a dag from a '!foreach' iterating a dag
A '!foreach' over a dag rebuilds the dag in place, applying its body to
the operator and to every argument, rather than collecting the body
values into a list. Its result is therefore the dag itself and the body
type carries no information about it, whereas the list case stays a
'list' of whatever the body yields.

The iteration variable follows the same split: an element type when
iterating a list, but a dag when iterating a dag. That mirrors TableGen,
which declares the variable with the dag type even though folding later
substitutes arbitrary argument values for it, and it is what decides
which programs are accepted.

The type of a dag '!foreach' hides its iteration variable, so the latter
is only observable through diagnostics. Cover it in the semantic
annotator's tests instead, which had no test class of its own yet; the
argument validation tests living in the syntax annotator's tests move
there as well, leaving each class with the diagnostics of one annotator.
